+++ docs/CHANGELOG.rst
@@ -32,6 +32,7 @@ Fixed
 - Fix pydocstyle warnings (PEP 257)
 - Take last version of process for spawned objects
 - Use default values for descriptor fields that are not given
+- Improve handling of validation errors
 
 ==================
 1.3.1 - 2016-07-27